requirements.txt
streamlit==1.35.0
sentence-transformers==3.0.1
pinecone==6.0.0
rank-bm25==0.2.2
huggingface-hub==0.23.4
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.33.0
